+++ src/controllers/issue-controller.ts
@@ -10,7 +10,10 @@ interface TimeEntryActivityItem extends vscode.QuickPickItem {
 }
 
 export class IssueController {
-  constructor(private issue: Issue, private redmine: RedmineServer) {}
+  constructor(
+    private issue: Issue,
+    private redmine: RedmineServer
+  ) {}
 
   chooseTimeEntryType(activities: TimeEntryActivity[]) {
     vscode.window
@@ -134,7 +137,7 @@ export class IssueController {
     let memberships: Membership[];
     try {
       memberships = await this.redmine.getMemberships(this.issue.project.id);
-    } catch (error) {
+    } catch (_error) {
       vscode.window.showErrorMessage(
         `Could not get memberships of project ${this.issue.project.name}`
       );
@@ -144,7 +147,7 @@ export class IssueController {
     let possibleStatuses: IssueStatus[];
     try {
       possibleStatuses = await this.redmine.getIssueStatusesTyped();
-    } catch (error) {
+    } catch (_error) {
       vscode.window.showErrorMessage("Could not get possible issue statuses");
       return;
     }